SDK Update,GP Services Update,GP Services fix, GP Services reworked buttons

+++ Assets/GooglePlayGames/Editor/GPGSUpgrader.cs
@@ -33,10 +33,10 @@ namespace GooglePlayGames.Editor
         static GPGSUpgrader()
         {
             string prevVer = GPGSProjectSettings.Instance.Get(GPGSUtil.LASTUPGRADEKEY, "00000");
-            if (prevVer != PluginVersion.VersionKey)
+            if (!prevVer.Equals(PluginVersion.VersionKey))
             {
                 // if this is a really old version, upgrade to 911 first, then 915
-                if (prevVer != PluginVersion.VersionKeyCPP)
+                if (!prevVer.Equals(PluginVersion.VersionKeyCPP))
                 {
                     prevVer = Upgrade911(prevVer);
                 }
@@ -52,8 +52,11 @@ namespace GooglePlayGames.Editor
 
                 prevVer = Upgrade931(prevVer);
 
+                prevVer = Upgrade935(prevVer);
+
                 // there is no migration needed to 930+
-                if (prevVer != PluginVersion.VersionKey) {
+                if (!prevVer.Equals(PluginVersion.VersionKey))
+                {
                     Debug.Log("Upgrading from format version " + prevVer + " to " + PluginVersion.VersionKey);
                     prevVer = PluginVersion.VersionKey;
                 }
@@ -101,7 +104,7 @@ namespace GooglePlayGames.Editor
             // Check that there is a AndroidManifest.xml file
             if (!GPGSUtil.AndroidManifestExists())
             {
-                GPGSUtil.GenerateAndroidManifest(false);
+                GPGSUtil.GenerateAndroidManifest();
             }
 
             AssetDatabase.Refresh();
@@ -136,6 +139,61 @@ namespace GooglePlayGames.Editor
             {
                 CleanDuplicates(s);
             }
+        }
+
+        /// <summary>
+        /// Upgrade to 0.9.35
+        /// </summary>
+        /// <remarks>
+        /// This cleans up some unused files mostly related to the improved jar resolver.
+        /// </remarks>
+        /// <param name="prevVer">Previous ver.</param>
+        private static string Upgrade935(string prevVer)
+        {
+            string[] obsoleteFiles =
+                {
+                "Assets/GooglePlayGames/Editor/CocoaPodHelper.cs",
+                "Assets/GooglePlayGames/Editor/CocoaPodHelper.cs.meta",
+                "Assets/GooglePlayGames/Editor/GPGSInstructionWindow.cs",
+                "Assets/GooglePlayGames/Editor/GPGSInstructionWindow.cs.meta",
+                "Assets/GooglePlayGames/Editor/Podfile.txt",
+                "Assets/GooglePlayGames/Editor/Podfile.txt.meta",
+                "Assets/GooglePlayGames/Editor/cocoapod_instructions",
+                "Assets/GooglePlayGames/Editor/cocoapod_instructions.meta",
+                "Assets/GooglePlayGames/Editor/ios_instructions",
+                "Assets/GooglePlayGames/Editor/ios_instructions.meta",
+
+                "Assets/PlayServicesResolver/Editor/DefaultResolver.cs",
+                "Assets/PlayServicesResolver/Editor/DefaultResolver.cs.meta",
+                "Assets/PlayServicesResolver/Editor/IResolver.cs",
+                "Assets/PlayServicesResolver/Editor/IResolver.cs.meta",
+                "Assets/PlayServicesResolver/Editor/JarResolverLib.dll",
+                "Assets/PlayServicesResolver/Editor/JarResolverLib.dll.meta",
+                "Assets/PlayServicesResolver/Editor/PlayServicesResolver.cs",
+                "Assets/PlayServicesResolver/Editor/PlayServicesResolver.cs.meta",
+                "Assets/PlayServicesResolver/Editor/ResolverVer1_1.cs",
+                "Assets/PlayServicesResolver/Editor/ResolverVer1_1.cs.meta",
+                "Assets/PlayServicesResolver/Editor/SampleDependencies.cs",
+                "Assets/PlayServicesResolver/Editor/SampleDependencies.cs.meta",
+                "Assets/PlayServicesResolver/Editor/SettingsDialog.cs",
+                "Assets/PlayServicesResolver/Editor/SettingsDialog.cs.meta",
+
+                "Assets/Plugins/Android/play-services-plus-8.4.0.aar",
+                "Assets/PlayServicesResolver/Editor/play-services-plus-8.4.0.aar.meta",
+
+                // not an obsolete file, but delete the cache since the schema changed.
+                "ProjectSettings/GoogleDependencyGooglePlayGames.xml"
+            };
+            foreach (string file in obsoleteFiles)
+            {
+                if (File.Exists(file))
+                {
+                    Debug.Log("Deleting obsolete file: " + file);
+                    File.Delete(file);
+                }
+            }
+
+            return PluginVersion.VersionKey;
         }
 
         /// <summary>
@@ -190,12 +248,17 @@ namespace GooglePlayGames.Editor
                 "Assets/GooglePlayGames/Platforms/Android/Gms/Games/Stats/StatsObject.cs.meta"
             };
 
-            foreach (string file in obsoleteFiles)
+            // only delete these if we are not version 0.9.34
+            if (string.Compare(PluginVersion.VersionKey, PluginVersion.VersionKeyJNIStats,
+                               System.StringComparison.Ordinal) <= 0)
             {
-                if (File.Exists(file))
+                foreach (string file in obsoleteFiles)
                 {
-                    Debug.Log("Deleting obsolete file: " + file);
-                    File.Delete(file);
+                    if (File.Exists(file))
+                    {
+                        Debug.Log("Deleting obsolete file: " + file);
+                        File.Delete(file);
+                    }
                 }
             }
